Changed variable names in Aplysia model.
Changed brain.mod per reviewer's suggestion. In manuscript, aIn1 and aIn2 were used with a0 and a1, respectively. This appeared to be inconsistent, though it's not wrong and actually won't match the other compute functions that use the pairs (a1,a2) and (a2,a0). However, it will at least look nicer in the example given in the paper.